Task/62 handle end game (#81)
* Basic handling of the end of a game

- Return GAME_OVER status when player has sufficient score
- Update UI to display a simple podium view
- Fix bug when concurrent pile hit (delete round in game manager)

* Persist score on game over

Don't remove the player from the database on game over,  so the score is kept forever.

* Improve podium component

* Remove react-use and BEMify css

+++ server/Services/GameService.cs
@@ -15,10 +15,10 @@ namespace Api.Services
     List<Game> GetActive();
     Game Get(int gameId);
     Player AddPlayer(int gameId, string playerId = null);
-    void RemovePlayer(string playerId);
+    void RemovePlayer(int gameId, string playerId);
     Game UpdateSettings(int gameId, Difficulty difficulty, TimeSpan cardSpeed, int roundsToWin);
     Game UpdateStatus(int gameId, GameStatus status);
-    void UpdatePlayerScore(string playerId, int value);
+    int? UpdatePlayerScore(string playerId, int value);
   }
 
   public class GameService : IGameService
@@ -88,8 +88,8 @@ namespace Api.Services
     {
       var game = _dbContext.Games.Find(gameId);
 
-      if (game.Status == GameStatus.ROUND_IN_PROGRESS)
-        throw new InvalidOperationException("You can't join a game when a round is in progress");
+      if (game.Status == GameStatus.GAME_OVER || game.Status == GameStatus.ROUND_IN_PROGRESS)
+        return null;
 
       var player = new Player();
       player.Id = playerId ?? Guid.NewGuid().ToString();
@@ -102,25 +102,32 @@ namespace Api.Services
       return player;
     }
 
-    public void RemovePlayer(string playerId)
+    public void RemovePlayer(int gameId, string playerId)
     {
-      var player = _dbContext.Players.Find(playerId);
+      var game = this.Get(gameId);
+
+      var player = game.Players.Where(player => player.Id == playerId).FirstOrDefault();
       if (player == null)
         throw new ArgumentException($"The playerId {playerId} was not found.");
-      _dbContext.Players.Remove(player);
-      _dbContext.SaveChanges();
+
+      // Keep player if game is over to persist score
+      if (game.Status != GameStatus.GAME_OVER)
+      {
+        _dbContext.Players.Remove(player);
+        _dbContext.SaveChanges();
+      }
     }
 
     // A player can win or lose points
-    // If the player can't be found, it means he has left the game: do nothing
-    public void UpdatePlayerScore(string playerId, int value)
+    // If the player can't be found, it means he has probably left the game: do nothing
+    public int? UpdatePlayerScore(string playerId, int value)
     {
       var player = _dbContext.Players.Find(playerId);
-      if (player != null)
-      {
-        player.Score = player.Score + value;
-        _dbContext.SaveChanges();
-      }
+      if (player == null)
+        return null;
+      player.Score = player.Score + value;
+      _dbContext.SaveChanges();
+      return player.Score;
     }
 
     private string GetRandomName()
